remove about user control and merge to view.
